Mesos containerizer can now launch Docker images.
When the Docker store is part of the container the layers cannot be untarred. The following error occurs:

`tar: etc/rc0.d/.wh.K01sendsigs: Cannot open: Operation not permitted`

Therefore a volume /var/lib/mesos/agent-$UUID is created under which the Docker store lives. When the agent is destroyed the volume is removed. The existing sandbox host volume maps to a subdirectory of the agent volume and will be empty when the cluster is destroyed.