Copy `minimesosFile` into the `minimesos-cli` container
Previously, the `minimesos-cli` container relied on `minimesosFile` present
in the host's $MINIMESOS_HOST_DIR mounted inside the container. This works
for a local docker daemon, but fails with remote docker daemons (e.g.
docker-machine), because $DOCKER_CONTAINER:$MINIMESOS_HOST_DIR/minimesosFile
is a fresh directory if it does not exist in the [remote] host file system.

Now `minimesosFile` is copied into $DOCKER_CONTAINER:/minimesosFile and
minimesos-cli is configured with a new system property that points to this
location.

This should fix https://github.com/ContainerSolutions/minimesos/issues/546